No need for cast
git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@24462 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/client/async/SplitFileFetcherSubSegment.java
@@ -170,7 +170,7 @@ public class SplitFileFetcherSubSegment extends SendableGet implements SupportsB
 				int x;
 				if(blockNums.size() == 0) return null;
 				x = context.random.nextInt(blockNums.size());
-				ret = (Integer) blockNums.remove(x);
+				ret = blockNums.remove(x);
 				Key key = segment.getBlockNodeKey(((Integer)ret).intValue(), container);
 				if(key == null) {
 					if(segment.isFinishing(container) || segment.isFinished(container)) return null;
